My last commit didn't have most of the changes staged, whoops, so here it is again.
"Adds new feature allowing user from a closed chapter to send an invite
to an email allowing new user to join closed chapter.

Removed add user button from manage users page and disabled add user
url.

Added invite user button to manage users and added invite user page.

A user can only invite members to their chapter. A staff member has the
option to give staff access to invitee. A superuser may give staff
and/or superuser access to invitee.

Invite generates a 256 bit join token and sends a link to invitee as and
email. Saves email, token, staff/superuser accesses, timestamp and
chapter url. Invites are only valid for 2 days after they are sent.
Old invites are not deleted or garbage cleaned or anything unless that
happens automatically.

There is a bug when loading the invite model from the database. It
switches all the values in the fields? so invite.email will be the token
value, invite.TOKEN will give you the superuser access flag?

Also when joining a chapter, it will not say any form is valid. So you
can't join anything.

If it weren't for those two errors, everything would work fine I think
but I can't exactly test it.

Also the psuedorandom number generation for the token is not
cryptographically secure yet as the seed is not securely generated. That
should be fixed in a later update. Also it calls the os's random number
generator so if our server doesn't have that the tokens can't generate
and it'll send people to and error page."